[Fix] layout.js 구조 수정
