+++ ssg_point_app_fe/app/(coupon)/couponPage/page.tsx
@@ -1,6 +1,11 @@
 import React from 'react'
 import Footer from '@/components/layout/Footer'
 
+export const metadata = {
+  title: '신세계포인트 - 쿠폰',
+  description: '쿠폰 페이지',
+}
+
 export default function page() {
   return (
     <div>
