+++ ssg_point_app_fe/app/(login)/login/page.tsx
@@ -3,6 +3,11 @@ import BottomNavBar from '@/components/layout/FootNav'
 
 import React from 'react'
 
+export const metadata = {
+  title: '신세계포인트 - 로그인',
+  description: '로그인 페이지',
+}
+
 export default function page() {
   return (
     <div>
