+++ ssg_point_app_fe/app/layout.tsx
@@ -1,5 +1,5 @@
 import Header from '@/components/layout/Header'
-import '../globals.css'
+import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import Footer from '@/components/layout/Footer'
